* EOControl/EONSAddOns.h/m:         ([NSString -parsedFirstVersionSubstring]): New method.         * EOAdaptors/Postgres95/Postgres95Adaptor.h/m: Added include of         pg_config.h to access PG_VERSION.  Added databaseVersion to list         of meaningful connectionDictionary keys.         (postgresClientVersion): New function.         * EOAdaptors/Postgres95/Postgres95Channel.h/m:  Added instance         variable to hold the version of the database server.         (pgResultDictionary): New static function for debuging.         ([Postgres95Channel -_readServerVersion]): New method to set Server         Version.         ([Postgres95Channel -openChannel]): Call _readServerVersion method.         ([Postgres95Channel -describeTableNames]): Adapt select statement         according to database version.         * Postgres95/Postgres95SQLExpression.m:         ([Postgres95SQLExpression +dropTableStatementsForEntityGroup:]):         Adapt select statement according to database version supplied in         connectionDictionary of the entites model.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@17013 72102866-910b-0410-8b05-ffd578937521

+++ EOAdaptors/Postgres95/Postgres95SQLExpression.m
@@ -44,10 +44,12 @@ RCS_ID("$Id$")
 #endif
 
 #include <EOControl/EONull.h>
+#include <EOControl/EONSAddOns.h>
 #include <EOControl/EODebug.h>
 
 #include <EOAccess/EOAttribute.h>
 #include <EOAccess/EOEntity.h>
+#include <EOAccess/EOModel.h>
 #include <EOAccess/EOSchemaGeneration.h>
 
 #include <Postgres95EOAdaptor/Postgres95SQLExpression.h>
@@ -305,18 +307,30 @@ RCS_ID("$Id$")
 
 + (NSArray *)dropTableStatementsForEntityGroup:(NSArray *)entityGroup
 {
-  // We redefine this method to add the CASCADE: it is needed to delete
-  // associated foreign key constraints when dropping a table.
-  // Q: shouldn't we move this into EOSQLExpression.m?
-  NSArray *newArray = nil;
+  EOEntity *entity;
+  NSArray  *newArray;
+  int       version;
 
   EOFLOGClassFnStartOrCond(@"EOSQLExpression");
 
-  newArray = [NSArray arrayWithObject:
-			[self expressionForString:
-				[NSString stringWithFormat: @"DROP TABLE %@ CASCADE",
-					  [[entityGroup objectAtIndex: 0]
-					    externalName]]]];
+  entity = [entityGroup objectAtIndex: 0];
+  version = [[[[entity model] connectionDictionary]
+	      objectForKey: @"databaseVersion"] parsedFirstVersionSubstring];
+  if (version == 0)
+    {
+      version = postgresClientVersion();
+    }
+
+  if (version < 70300)
+    {
+      newArray = [super dropTableStatementsForEntityGroup: entityGroup];
+    }
+  else
+    {
+      newArray = [NSArray arrayWithObject: [self expressionForString:
+	[NSString stringWithFormat: @"DROP TABLE %@ CASCADE",
+		  [entity externalName]]]];
+    }
 
   EOFLOGClassFnStopOrCond(@"EOSQLExpression");
